Fix bug where all-zero message is not published
When "quiet mode" is on, we don't publish any message when all values in the
Twist message is zero (i.e. when there is no input from the joypad). When
there is no input to the joypad, we need to publish a message with all values
being zero in order to stop the manipulator.