Add action definition feature
This feature is built on top of the "recovery" action. It works by having the
user define presets, which as a set of "tasks" executed on call. Each preset
can have a mapping that triggers it, defined in the ~mapping/presets ROS
parameter. The tasks that are currently available are:

- move_group_state: Uses move_group to go to a certain pre-defined joint
  state.
- remember_move_group_state: Remembers the current joint state in a certain
  name so that it can later be recalled.
- sleep: Simply sleeps for a certain duration.
- single_point_trajectory: Publishes a JointTrajectory message containing one
  JointTrajectoryPoint.

The presets can be triggered using the following events:

- button_down: Triggered when the button is pressed down
- button_up: Triggered when the button is released
- double_press: Triggered when the button is pressed twice within 0.25 seconds
- triple_press: Triggered when the button is pressed three times within 0.25
  seconds
- long_press: Triggered when the button is pressed down over a certain time
  period (the time period is defined using the duration parameter)

Check out the YAML files found in the config directory for examples.