Make generic configurations possible
There are too many changes to the program that it is nearly impossible to
commit every change. Following is a brief list of the modifications and
improvements made to the program:

Generic actions
---------------
Previously, twist messages for the arm was treated differently from all the
other presets such as moving to a move_group state. Now, publishing twist
messages (not just for the arm but anything) is simply a preset which is
triggered for every Joy message that is received.

Generic axis and button description
-----------------------------------
Previously, physical buttons and axes were bound to presets. In the new
version, virtual axes and buttons are introduced in order to simplify the
preset "triggers." This is done by giving a name to a physical axis/button and
using that name to defined the trigger. With this new feature, one can use
physical axes as buttons (an example of this is the cursor buttons) and vice
versa.

Shared variable
---------------
Previously, it was impossible to share variables among "actions" ("preset
task" in the previous terminology). With the shared variable, we can, for
example, enable certain presets in a round-robin fashion.

Command triggers
----------------
Previously, presets could only be bound to either a single press, double
press, triple press, or a long press. No simultaneous press was allowed, and
multiple presses could be used. In the new version, arbitrary commands can be
defined as triggers, such as UUDDLRLRXO[1] or pressing R1, L1, R2, L2 at once.

[1] https://en.wikipedia.org/wiki/Konami_Code